// ddg-frontend/src/services/openapi/models/PdfCompanyTutor_pdf_write.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PdfCompanyTutor_pdf_write = {
    name?: string;
    lastName?: string;
    email?: string | null;
    phone?: string | null;
    nif?: string | null;
    address?: string | null;
    city?: string | null;
    postalCode?: string | null;
};